requirements.txt
python-dotenv
llama-index
pypdf
